Making simplifiers configurable too
