+++ spark_board/plan_extractor/dag_builder.py
@@ -2,25 +2,26 @@ from py4j.java_gateway import JavaObject
 from pyspark.sql import DataFrame
 from typing import List, Dict, Any, Generator, Optional
 
-from .dag_simplification import simplify
+from .dag_simplification import simplify, default_simplifiers
+from .dag_simplification.dag_simplifier import DagSimplifier
 from .transformations_dag import TransformationNode
 from .transformation_node_builders import build_dag_from_java_object
-from .transformation_node_builders.repository import Repository, create_default
+from .transformation_node_builders.repository import Repository as TnbRepository, create_default
 from .py4j_utils import iterate_java_object
 
 
-def build_dag(df: DataFrame, simplify_dag: bool = True, repository: Repository=create_default()) -> TransformationNode:
+def build_dag(df: DataFrame, dag_simplifiers: List[DagSimplifier] = default_simplifiers(), tnb_repository: TnbRepository=create_default()) -> TransformationNode:
     """
     Builds a DAG from a PySpark's DataFrame.
     
-    :param    df:                             The PySpark's DataFrame
-    :param    simplify_dag:                   Indicates whether the DAG simplifiers should be applied or not.
-    :param    repository:                     A repository of specific transformation node builders.    
+    :param    df:               The PySpark's DataFrame.
+    :param    dag_simplifiers:  The list of DAG simplifiers to be applied.
+    :param    tnb_repository:   A repository of specific transformation node builders.
     :returns: The DAG for the passed PySpark's DataFrame.
     """
     first_child = _get_last_transformation(df)
-    raw_dag = build_dag_from_java_object(node=first_child, repository=repository)
-    return simplify(raw_dag) if simplify_dag else raw_dag
+    raw_dag = build_dag_from_java_object(node=first_child, repository=tnb_repository)
+    return simplify(raw_dag, dag_simplifiers)
 
 
 def _get_last_transformation(df: DataFrame) -> JavaObject:
