+++ spark_board/plan_extractor/dag_simplification/merge_join_and_project.py
@@ -1,23 +1,8 @@
-from .transformations_dag import Metadata, TransformationColumn, TransformationNode, TransformationType, JoinCondition
-
 from typing import Dict, List, Set, TypeVar
 
+from ..transformations_dag import Metadata, TransformationColumn, TransformationNode, TransformationType, JoinCondition
 
-class DagSimplifier(object):
-    """
-    The DAG Simplifiers detect certain pattern in a DAG, and map them to
-    a simplified version of it.
-    For example, four filters in a row could be simplified as one filter
-    with four conditions.
-    """
-
-    def can_apply(self, dag: TransformationNode) -> bool:
-        """Determines whether a pattern can be simplified by this simplifier or not"""
-        raise NotImplementedError("Abstract method")
-    
-    def apply(self, original: TransformationNode) -> TransformationNode:
-        """Returns a simplified version of the original DAG"""
-        raise NotImplementedError("Abstract method")
+from .dag_simplifier import DagSimplifier
 
 
 class MergeJoinAndProject(DagSimplifier):
@@ -97,69 +82,7 @@ class MergeJoinAndProject(DagSimplifier):
 
         return pcols
 
-class MergeDatasourceAndAlias(DagSimplifier):
-    """
-    When the PySpark dataframe represents a named table, it generates
-    two nodes: one containing the table data itself, and another one
-    containing an alias. This simplifier unifies both adding an
-    attribute to the table node.
-    When the datasource is a logical RDD, the alias is not added, so
-    this simplifier considers that case and skips it.
-    """
-
-    def __init__(self) -> None:
-        self.already_merged_relations: Set[TransformationNode] = set()
-
-    def can_apply(self, df: TransformationNode) -> bool:
-        if df.type != TransformationType.Alias:
-            return False
-        alias_node = df
-        if alias_node in self.already_merged_relations:
-            return False
-        if len(alias_node.children) != 1:
-            return False
-        relation_node = alias_node.children[0]
-        if relation_node.type != TransformationType.Relation:
-            return False
-        return True
-
-    def apply(self, original: TransformationNode) -> TransformationNode:
-        alias_node = original
-        relation_node = original.children[0]
-        new_node = TransformationNode(
-            type=TransformationType.Relation,
-            metadata=relation_node.metadata,
-            children=relation_node.children,
-            columns=self._remove_links(alias_node.columns),
-        )
-        for c in new_node.columns.values():
-            c.node_id = new_node.id
-        self.already_merged_relations.add(new_node)
-        return new_node
-
-    def _remove_links(self, alias_cols: Dict[int, TransformationColumn]) -> Dict[int, TransformationColumn]:
-        for col in alias_cols.values():
-            col.links = []
-        return alias_cols
-
 
 T = TypeVar("T")
 def flatten(l: List[List[T]]) -> List[T]:
     return [item for sublist in l for item in sublist]
-
-
-all_heuristics = [
-    MergeJoinAndProject(),
-    MergeDatasourceAndAlias(),
-]
-
-
-def simplify(raw_dag: TransformationNode) -> TransformationNode:
-    # heuristics are applied with a "child-first" policy
-    raw_dag.children = [simplify(child) for child in raw_dag.children]
-
-    result_dag = raw_dag
-    for heuristic in all_heuristics:
-        if heuristic.can_apply(result_dag):
-            result_dag = heuristic.apply(result_dag)
-    return result_dag
